add cli
it works like parse example but with extended functionality
