+++ Cargo.toml
@@ -5,7 +5,9 @@ edition = "2024"
 
 [dependencies]
 bumpalo = { version = "3.19.0", features = ["collections", "boxed"] }
-
-
-[dev-dependencies]
+clap = { version = "4.5.20", features = ["derive"] }
 ptree = { version = "0.5.2", features = ["ansi"] }
+
+[[bin]]
+name = "cofy"
+path = "src/main.rs"
