+++ README.md
@@ -1,3 +1,4 @@
 # cofy
 Cofy is a WIP interpreted, functional & embeddable programming language. Check `lib/` or `test/` for code snippets.
-If you want to play around with the parser of the language run `cargo r --example parse`.
+If you want to play around with the parser of the language run `cargo r`. You can also pass path to a file or script
+directly via `cargo r -- /path/to/cofy/script.cofy` or `cargo r -- -c '<script>'`.
