+++ src/ast_rename.rs
@@ -1,14 +1,9 @@
-use bumpalo::Bump;
 use cofy::{
-    Expr, ExprInner, Lexer, LexerError, Operator, Parser,
+    Expr, ExprInner, Operator, Parser,
     parser::{AstInner, Type},
 };
 use ptree::TreeBuilder;
-use std::{
-    error::Error,
-    io::{Error as IoErr, stdin},
-    ops::Deref,
-};
+use std::ops::Deref;
 
 fn op_to_str(op: Operator) -> &'static str {
     match op {
@@ -109,7 +104,7 @@ fn print_type(tree: &mut TreeBuilder, ty: &Type) {
     };
 }
 
-fn print_tree(ast: Parser) {
+pub fn print_tree(ast: Parser) {
     let mut tree = TreeBuilder::new("stdin's abstract syntax tree".to_owned());
     for a in ast {
         match a {
@@ -162,29 +157,4 @@ fn print_tree(ast: Parser) {
 
     let tree = tree.build();
     ptree::print_tree(&tree).expect("cannot print the tree")
-}
-
-fn main() -> Result<(), Box<dyn Error>> {
-    let bump = Bump::new();
-    println!("write your code here");
-    println!("print `end` to start the parser");
-    println!();
-
-    let input = stdin()
-        .lines()
-        .take_while(|a| match a.as_deref() {
-            Ok("end") | Err(_) => false,
-            _ => true,
-        })
-        .collect::<Result<Vec<String>, IoErr>>()?
-        .join("\n");
-
-    println!();
-
-    let lexer = Lexer::new(&input);
-    let tokens = lexer.collect::<Result<Vec<_>, LexerError>>()?;
-    print_tree(Parser::new(&bump, &tokens));
-    println!("bytes allocated on the bump: {}", bump.allocated_bytes());
-
-    Ok(())
 }
